# server/src/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CricZone API
  version: 1.0.0
  description: |
    REST API for CricZone cricket analytics.

    The API provides access to matches, scorecards, teams, players,
    series, venues, rankings, news, and cricket statistics.

    Live incremental match updates are delivered separately through
    Socket.IO and are documented in docs/events/websocket-events.md.
servers:
  - url: /api/v1
    description: CricZone API v1

# CricZone public browsing API.
# No authentication is required in Phase 1.
security: []
...
paths:
...
  /venues:
    get:
      tags: [Venues]
      summary: Get all venues
      operationId: getVenues
      responses:
        "200":
          description: Venues retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /venues/{venueId}/matches:
    get:
      tags: [Venues]
      summary: Get matches played at a venue
      operationId: getMatchesByVenueId
      parameters:
        - $ref: "#/components/parameters/VenueId"
      responses:
        "200":
          description: Venue matches retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /rankings:
    get:
      tags: [Rankings]
      summary: Get all rankings
      operationId: getRankings
      responses:
        "200":
          description: Rankings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /rankings/players:
    get:
      tags: [Rankings]
      summary: Get player rankings
      operationId: getPlayerRankings
      responses:
        "200":
          description: Player rankings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /rankings/teams:
    get:
      tags: [Rankings]
      summary: Get team rankings
      operationId: getTeamRankings
      responses:
        "200":
          description: Team rankings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
...
    VenueId:
      name: venueId
      in: path
      required: true
      schema:
        type: integer
      description: Venue identifier
...
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Resource was not found
            requestId:
              type: string
              nullable: true
    DataResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Endpoint-specific response data
    Match:
      type: object
      properties:
        id:
          type: integer
        format:
          type: string
        status:
          type: string
        start_time:
          type: string
          format: date-time
        result:
          type: string
          nullable: true
        series_id:
          type: integer
        series_name:
          type: string
        venue_id:
          type: integer
        venue_name:
          type: string
        venue_city:
          type: string
        team1_id:
          type: integer
        team1_name:
          type: string
        team1_short_name:
          type: string
        team2_id:
          type: integer
        team2_name:
          type: string
        team2_short_name:
          type: string
...
    MatchListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Match"
...
    Venue:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    VenueListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
...